code refactpring
