+++ src/lib/services/assets/assets-service.ts
@@ -1,11 +1,5 @@
-import type { IAssetInfoService } from '.';
+import type { AssetInfo, IAssetInfoService } from '.';
 import type { INodeClient } from '$lib/services/node-client';
-
-export type AssetInfo = {
-  decimals: number;
-  assetId: string;
-  symbol: string;
-};
 
 export class AssetService implements IAssetInfoService {
   nodeClient: INodeClient;
