+++ src/lib/services/assets/index.ts
@@ -1,5 +1,11 @@
-import type { AssetInfo } from './assets-service';
-
 export type IAssetInfoService = {
   getAssetInfo(assetId: string): Promise<AssetInfo>;
 };
+
+export type AssetInfo = {
+  decimals: number;
+  assetId: string;
+  symbol: string;
+};
+
+export { AssetService } from './assets-service';
