+++ src/lib/services/balance/index.ts
@@ -1,8 +1,15 @@
 import type { INodeClient } from '$lib/services/node-client';
-import type { BalanceType } from './balance-service';
 
 export type IBalanceService = {
   nodeClient: INodeClient;
   address: string;
   subscribe(balanceType: BalanceType, assetId: string): ReadableStreamDefaultReader<bigint>;
 };
+
+export enum BalanceType {
+  wallet = 'wallet',
+  inOrders = 'inOrders',
+  lockedInOrders = 'lockedInOrders'
+}
+
+export { BalanceService } from './balance-service';
