+++ src/lib/stores/balance.ts
@@ -1,7 +1,7 @@
 import type { INodeClient } from '$lib/services/node-client';
 import { readable } from 'svelte/store';
-import { WavesHttpNodeClient } from 'src/lib/services/node-client/node-clients';
-import { BalanceService, BalanceType } from '$lib/services/balance/balance-service';
+import { WavesHttpNodeClient } from 'src/lib/services/node-client/waves-http-node-client';
+import { BalanceService, BalanceType } from '$lib/services/balance';
 import { AssetIds, WAVES_NODES_BASE_URL } from '../constants';
 
 const address: string = 'address';
